requirements-mac.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fsspec==2024.6.1
groq==0.9.0
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
identify==2.6.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
llama_cpp_python==0.2.87
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
meson==1.5.1
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
ninja==1.11.1.1
nodeenv==1.9.1
numpy==1.26.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.4
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.44.0
yarg==0.1.9
